+++ lib/generators/rodauth/templates/db/migrate/create_rodauth.rb
@@ -4,7 +4,7 @@ class CreateRodauth < ActiveRecord::Migration<%= migration_version %>
     enable_extension "citext"
 
 <% end -%>
-    create_table :accounts do |t|
+    create_table :accounts<%= primary_key_type %> do |t|
 <% case activerecord_adapter -%>
 <% when "postgresql" -%>
       t.citext :email, null: false, index: { unique: true, where: "status IN ('verified', 'unverified')" }
@@ -15,13 +15,13 @@ class CreateRodauth < ActiveRecord::Migration<%= migration_version %>
     end
 
     # Used if storing password hashes in a separate table (default)
-    create_table :account_password_hashes do |t|
+    create_table :account_password_hashes<%= primary_key_type %> do |t|
       t.foreign_key :accounts, column: :id
       t.string :password_hash, null: false
     end
 
     # Used by the password reset feature
-    create_table :account_password_reset_keys do |t|
+    create_table :account_password_reset_keys<%= primary_key_type %> do |t|
       t.foreign_key :accounts, column: :id
       t.string :key, null: false
       t.datetime :deadline, null: false
@@ -29,7 +29,7 @@ class CreateRodauth < ActiveRecord::Migration<%= migration_version %>
     end
 
     # Used by the account verification feature
-    create_table :account_verification_keys do |t|
+    create_table :account_verification_keys<%= primary_key_type %> do |t|
       t.foreign_key :accounts, column: :id
       t.string :key, null: false
       t.datetime :requested_at, null: false, default: -> { "CURRENT_TIMESTAMP" }
@@ -37,7 +37,7 @@ class CreateRodauth < ActiveRecord::Migration<%= migration_version %>
     end
 
     # Used by the verify login change feature
-    create_table :account_login_change_keys do |t|
+    create_table :account_login_change_keys<%= primary_key_type %> do |t|
       t.foreign_key :accounts, column: :id
       t.string :key, null: false
       t.string :login, null: false
@@ -46,14 +46,14 @@ class CreateRodauth < ActiveRecord::Migration<%= migration_version %>
 
 <% unless api_only? -%>
     # Used by the remember me feature
-    create_table :account_remember_keys do |t|
+    create_table :account_remember_keys<%= primary_key_type %> do |t|
       t.foreign_key :accounts, column: :id
       t.string :key, null: false
       t.datetime :deadline, null: false
     end
 <% else -%>
     # # Used by the remember me feature
-    # create_table :account_remember_keys do |t|
+    # create_table :account_remember_keys<%= primary_key_type %> do |t|
     #   t.foreign_key :accounts, column: :id
     #   t.string :key, null: false
     #   t.datetime :deadline, null: false
@@ -61,8 +61,8 @@ class CreateRodauth < ActiveRecord::Migration<%= migration_version %>
 <% end -%>
 
     # # Used by the audit logging feature
-    # create_table :account_authentication_audit_logs do |t|
-    #   t.references :account, foreign_key: true, null: false
+    # create_table :account_authentication_audit_logs<%= primary_key_type %> do |t|
+    #   t.references :account, foreign_key: true, null: false<%= primary_key_type(:type) %>
     #   t.datetime :at, null: false, default: -> { "CURRENT_TIMESTAMP" }
     #   t.text :message, null: false
 <% case activerecord_adapter -%>
@@ -78,8 +78,8 @@ class CreateRodauth < ActiveRecord::Migration<%= migration_version %>
     # end
 
     # # Used by the jwt refresh feature
-    # create_table :account_jwt_refresh_keys do |t|
-    #   t.references :account, foreign_key: true, null: false
+    # create_table :account_jwt_refresh_keys<%= primary_key_type %> do |t|
+    #   t.references :account, foreign_key: true, null: false<%= primary_key_type(:type) %>
     #   t.string :key, null: false
     #   t.datetime :deadline, null: false
     #   t.index :account_id, name: "account_jwt_rk_account_id_idx"
@@ -87,16 +87,16 @@ class CreateRodauth < ActiveRecord::Migration<%= migration_version %>
 
     # # Used by the disallow_password_reuse feature
     # create_table :account_previous_password_hashes do |t|
-    #   t.references :account, foreign_key: true
+    #   t.references :account, foreign_key: true<%= primary_key_type(:type) %>
     #   t.string :password_hash, null: false
     # end
 
     # # Used by the lockout feature
-    # create_table :account_login_failures do |t|
+    # create_table :account_login_failures<%= primary_key_type %> do |t|
     #   t.foreign_key :accounts, column: :id
     #   t.integer :number, null: false, default: 1
     # end
-    # create_table :account_lockouts do |t|
+    # create_table :account_lockouts<%= primary_key_type %> do |t|
     #   t.foreign_key :accounts, column: :id
     #   t.string :key, null: false
     #   t.datetime :deadline, null: false
@@ -104,7 +104,7 @@ class CreateRodauth < ActiveRecord::Migration<%= migration_version %>
     # end
 
     # # Used by the email auth feature
-    # create_table :account_email_auth_keys do |t|
+    # create_table :account_email_auth_keys<%= primary_key_type %> do |t|
     #   t.foreign_key :accounts, column: :id
     #   t.string :key, null: false
     #   t.datetime :deadline, null: false
@@ -112,13 +112,13 @@ class CreateRodauth < ActiveRecord::Migration<%= migration_version %>
     # end
 
     # # Used by the password expiration feature
-    # create_table :account_password_change_times do |t|
+    # create_table :account_password_change_times<%= primary_key_type %> do |t|
     #   t.foreign_key :accounts, column: :id
     #   t.datetime :changed_at, null: false, default: -> { "CURRENT_TIMESTAMP" }
     # end
 
     # # Used by the account expiration feature
-    # create_table :account_activity_times do |t|
+    # create_table :account_activity_times<%= primary_key_type %> do |t|
     #   t.foreign_key :accounts, column: :id
     #   t.datetime :last_activity_at, null: false
     #   t.datetime :last_login_at, null: false
@@ -126,26 +126,26 @@ class CreateRodauth < ActiveRecord::Migration<%= migration_version %>
     # end
 
     # # Used by the single session feature
-    # create_table :account_session_keys do |t|
+    # create_table :account_session_keys<%= primary_key_type %> do |t|
     #   t.foreign_key :accounts, column: :id
     #   t.string :key, null: false
     # end
 
     # # Used by the active sessions feature
     # create_table :account_active_session_keys, primary_key: [:account_id, :session_id] do |t|
-    #   t.references :account, foreign_key: true
+    #   t.references :account, foreign_key: true<%= primary_key_type(:type) %>
     #   t.string :session_id
     #   t.datetime :created_at, null: false, default: -> { "CURRENT_TIMESTAMP" }
     #   t.datetime :last_use, null: false, default: -> { "CURRENT_TIMESTAMP" }
     # end
 
     # # Used by the webauthn feature
-    # create_table :account_webauthn_user_ids do |t|
+    # create_table :account_webauthn_user_ids<%= primary_key_type %> do |t|
     #   t.foreign_key :accounts, column: :id
     #   t.string :webauthn_id, null: false
     # end
     # create_table :account_webauthn_keys, primary_key: [:account_id, :webauthn_id] do |t|
-    #   t.references :account, foreign_key: true
+    #   t.references :account, foreign_key: true<%= primary_key_type(:type) %>
     #   t.string :webauthn_id
     #   t.string :public_key, null: false
     #   t.integer :sign_count, null: false
@@ -153,7 +153,7 @@ class CreateRodauth < ActiveRecord::Migration<%= migration_version %>
     # end
 
     # # Used by the otp feature
-    # create_table :account_otp_keys do |t|
+    # create_table :account_otp_keys<%= primary_key_type %> do |t|
     #   t.foreign_key :accounts, column: :id
     #   t.string :key, null: false
     #   t.integer :num_failures, null: false, default: 0
@@ -162,13 +162,13 @@ class CreateRodauth < ActiveRecord::Migration<%= migration_version %>
 
     # # Used by the recovery codes feature
     # create_table :account_recovery_codes, primary_key: [:id, :code] do |t|
-    #   t.integer :id
+    #   t.column :id, :<%= primary_key_type(nil) || :bigint %>
     #   t.foreign_key :accounts, column: :id
     #   t.string :code
     # end
 
     # # Used by the sms codes feature
-    # create_table :account_sms_codes do |t|
+    # create_table :account_sms_codes<%= primary_key_type %> do |t|
     #   t.foreign_key :accounts, column: :id
     #   t.string :phone_number, null: false
     #   t.integer :num_failures